Unify two command handlers.
