+++ bots/pr/src/main/java/org/openjdk/skara/bots/pr/CommitCommandWorkItem.java
@@ -44,35 +44,6 @@ public class CommitCommandWorkItem implements WorkItem {
 
     private static final Logger log = Logger.getLogger("org.openjdk.skara.bots.pr");
 
-    static final Map<String, CommandHandler> commandHandlers = Map.ofEntries(
-            Map.entry("help", new HelpCommand()),
-            Map.entry("backport", new BackportCommand()),
-            Map.entry("tag", new TagCommand())
-    );
-
-    static class HelpCommand implements CommandHandler {
-        @Override
-        public void handle(PullRequestBot bot, HostedCommit commit, CensusInstance censusInstance, Path scratchPath, CommandInvocation command, List<Comment> allComments, PrintWriter reply) {
-            reply.println("Available commands:");
-            Stream.concat(
-                    commandHandlers.entrySet().stream()
-                                   .map(entry -> entry.getKey() + " - " + entry.getValue().description()),
-                    bot.externalCommitCommands().entrySet().stream()
-                       .map(entry -> entry.getKey() + " - " + entry.getValue())
-            ).sorted().forEachOrdered(c -> reply.println(" * " + c));
-        }
-
-        @Override
-        public String description() {
-            return "shows this text";
-        }
-
-        @Override
-        public boolean allowedInCommit() {
-            return true;
-        }
-    }
-
     CommitCommandWorkItem(PullRequestBot bot, CommitComment commitComment, Consumer<RuntimeException> onError) {
         this.bot = bot;
         this.commitComment = commitComment;
@@ -95,7 +66,7 @@ public class CommitCommandWorkItem implements WorkItem {
 
     private Optional<CommandInvocation> nextCommand(List<CommitComment> allComments) {
         var self = bot.repo().forge().currentUser();
-        var command = CommandExtractor.extractCommands(true, commitComment.body(),
+        var command = CommandExtractor.extractCommands(commitComment.body(),
                                                        commitComment.id(), commitComment.author());
         if (command.isEmpty()) {
             return Optional.empty();
